+++ CHANGES.md
@@ -1,8 +1,15 @@
 # ChangeLog
 
 ## v24.1.0
+Enhancements:
+* Improve graphs window #335
+* Option to edit the app fontsize in settings #335
+
 Bug fixes:
 * use eqt FormDialog in settings #318
+
+Dependencies:
+* Add pandas #335
 
 ## v24.0.1
 Bug fixes: